requirements.txt
roboflow
ultralytics
streamlit
numpy
pillow
opencv-python-headless